+++ src/session/session_itp.mli
@@ -248,6 +248,10 @@ val mark_obsolete: session -> proofAttemptID -> unit
 val save_session : session -> unit
 (** [save_session s] Save the session [s] *)
 
+val export_as_zip : session -> string
+(** [export_as_zip s] produces a zip archive of the session
+   file. Returns the file name of the archive. *)
+
 val load_session : string -> session
 (** [load_session dir] load a session in directory [dir]; all the
     tasks are initialised to None